Update menu label from 'C-Loader' to 'Component'
Changed the label for the 'component-loader' menu item to "Component" for improved clarity. The icon and template were not modified.

Update component label and increase UI element width

Changed the menu label from "C-Loader" to "Component" for clarity. Increased the width of the related UI element from 60px to 70px to improve appearance and accommodate the new label.